+++ src/main/java/domain/piece/TeamType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 public enum TeamType {
     CHO,
