+++ src/main/java/domain/piece/strategy/OuterElephantStrategy.java
@@ -1,7 +1,7 @@
 package domain.piece.strategy;
 
-import domain.Position;
-import domain.TeamType;
+import domain.position.Position;
+import domain.piece.TeamType;
 import domain.piece.Elephant;
 import domain.piece.Horse;
 import java.util.Map;
