+++ src/main/java/domain/position/Column.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.position;
 
 import java.util.Objects;
