+++ src/main/java/domain/position/Position.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.position;
 
 import java.util.Objects;
