+++ src/main/java/domain/position/Row.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.position;
 
 import java.util.Objects;
