+++ src/test/java/domain/BoardTest.java
@@ -9,10 +9,9 @@ import domain.piece.King;
 import domain.piece.Piece;
 import domain.piece.PieceType;
 import domain.piece.Soldier;
-import fixtures.PositionFixture;
-import java.util.List;
+import domain.piece.TeamType;
+import domain.position.Position;
 import java.util.Map;
-import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
