+++ src/test/java/domain/MovementTest.java
@@ -4,7 +4,9 @@ import static fixtures.PositionFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import fixtures.PositionFixture;
+import domain.position.Direction;
+import domain.position.Movement;
+import domain.position.Position;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
