+++ src/test/java/domain/RowTest.java
@@ -3,6 +3,7 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.position.Row;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
