+++ src/test/java/domain/piece/HorseTest.java
@@ -4,13 +4,9 @@ import static fixtures.PositionFixture.*;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.Position;
-import domain.TeamType;
-import fixtures.PositionFixture;
-import java.util.List;
+import domain.position.Position;
 import java.util.Map;
 import java.util.stream.Stream;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
