+++ src/test/java/domain/piece/path/CannonPathValidatorTest.java
@@ -6,8 +6,8 @@ import static fixtures.PositionFixture.E8;
 import static fixtures.PositionFixture.E9;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.Position;
-import domain.TeamType;
+import domain.position.Position;
+import domain.piece.TeamType;
 import domain.piece.Cannon;
 import domain.piece.Elephant;
 import domain.piece.Horse;
