+++ src/test/java/domain/piece/strategy/OuterElephantStrategyTest.java
@@ -6,16 +6,13 @@ import static fixtures.PositionFixture.G0;
 import static fixtures.PositionFixture.H0;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.Position;
-import domain.TeamType;
-import domain.piece.Elephant;
-import domain.piece.Horse;
+import domain.position.Position;
+import domain.piece.TeamType;
 import domain.piece.Piece;
 import domain.piece.PieceType;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
